correct old null

+++ src/CoreBundle/Controller/UserController.php
@@ -259,8 +259,7 @@ class UserController extends BaseController
         $form->submit($request->request->all(), $clearMissing);
 
         if ($form->isValid()) {
-            if ($user->getProfilePicture() != $oldProfilePicture) {
-                return \FOS\RestBundle\View\View::create(['old' => $oldProfilePicture, "basename" => basename($oldProfilePicture)], Response::HTTP_OK);
+            if ($user->getProfilePicture() != $oldProfilePicture && $oldProfilePicture != null) {
                 unlink(__DIR__.'/../../../web/uploads/'.basename($oldProfilePicture));
             }
             $em = $this->get('doctrine.orm.entity_manager');
